% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{column_mean}
\alias{column_mean}
\title{Computes column mean using rcpp}
\usage{
column_mean(x)
}
\arguments{
\item{x}{n-sample by p-feature matrix}
}
\value{
p-vector of columnwise mean
}
\description{
Computes column mean using rcpp
}
