% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{column_median}
\alias{column_median}
\title{Computes column median using rcpp}
\usage{
column_median(x)
}
\arguments{
\item{x}{n-sample by p-feature matrix}
}
\value{
p-vector of columnwise medians
}
\description{
Computes column median using rcpp
}
