% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{column_medianVar}
\alias{column_medianVar}
\title{Computes column variance based on supplied median/mean vectors}
\usage{
column_medianVar(x, m)
}
\arguments{
\item{x}{n-sample by p-feature matrix}

\item{m}{p-vector of column medians to compute variance from}
}
\value{
p-vecor of columnwise variance based on supplied median value
}
\description{
Computes column variance based on supplied median/mean vectors
}
