% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{K_S}
\alias{K_S}
\title{Computes columnwise K-S statistic}
\usage{
K_S(mt, mt2)
}
\arguments{
\item{mt}{n-sample by p-feature matrix for class1}

\item{mt2}{n-sample by p-feature matrix for class2}
}
\value{
p-vector of K-S statistics
}
\description{
Computes columnwise K-S statistic
}
