% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxSpanTree.R
\name{mstAll}
\alias{mstAll}
\title{Computes the maximum spanning tree (MST) of a set of log-ratios}
\usage{
mstAll(featMatrix, dcvRanking)
}
\arguments{
\item{featMatrix}{a n-sample by p-logratio matrix}

\item{dcvRanking}{a p-row log-ratio scoring matrix usually from dcvScore()$lrs}
}
\value{
A data.frame with the MST selected log-ratio features
use_pipe(export = TRUE)
}
\description{
Uses scoring from dcvScores() to construct and prune a log-ratio graph witht the maxi= spanning tree method
}
\examples{
mstAll()

}
\references{
Hinton, A.L., Mucha, P.J., (2021). Simultaneous variable selection and group association testing in sparse high dimensional omics data. XXXX.
}
