% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcvScoring.R
\name{dcvScores}
\alias{dcvScores}
\title{Computes the Differential Compositional Variation (DCV) Scoring ALgoritmn}
\usage{
dcvScores(
  logRatioMatrix,
  includeInfoGain = T,
  nfolds = 5,
  numRepeats = 1,
  seed_ = 8272008,
  rankOrder = T
)
}
\arguments{
\item{logRatioMatrix}{a n-sample by p-logratio matrix}

\item{includeInfoGain}{should the infromation gain score be added}

\item{nfolds}{number of parition used to compute and average scores}

\item{numRepeats}{number of repeats on the nfold paritions}

\item{seed_}{random seed control}

\item{rankOrder}{should the score be rank ordered by score along with computing the running number of distinct features (Adds non-trivial computational time).}
}
\value{
A matrix with multiplicatively imputed zeroes
use_pipe(export = TRUE)
}
\description{
Scores log-ration using the DCV metric (Hinton (2021))
}
\examples{
finalDCV()

}
\references{
Hinton, A.L., Mucha, P.J., (2021). Simultaneous variable selection and group association testing in sparse high dimensional omics data. XXXX.
}
