\name{diffCompVarRcpp-package}
\alias{diffCompVarRcpp-package}
\alias{diffCompVarRcpp}
\docType{package}
\title{
\packageTitle{diffCompVarRcpp}
}
\description{
\packageDescription{diffCompVarRcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{diffCompVarRcpp}
\packageIndices{diffCompVarRcpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{diffCompVarRcpp}

Maintainer: \packageMaintainer{diffCompVarRcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
